backend/requirements.txt
fastapi==0.128.8
httpx==0.28.1
pydantic==2.12.3
python-dotenv==1.1.1
python-multipart==0.0.20
qdrant-client==1.16.1
reportlab==4.4.10
uvicorn[standard]==0.39.0
pytest==8.4.2
